Moved user-profile to pages/user-profile
