+++ src/App.js
@@ -9,7 +9,7 @@ import Dashboard from "./components/dashboard/Dashboard";
 import Navigation from "./components/navigation/Navigation";
 import Footer from "./components/navigation/Footer";
 import Map from "./components/pages/comparison/Map";
-import Profile from "./components/user-profile/Profile";
+import Profile from "./components/pages/user-profile/Profile";
 import PrivacyPolicy from "./components/legal/PrivacyPolicy";
 import AboutUs from "./components/pages/aboutus/AboutUs";
 // import AboutUs2 from './components/aboutus/AboutUs2';
