+++ src/components/pages/user-profile/icons/checkmark.png
